house_hearing_excerpt dataset

Sample text dataset used for demonstration of `corporations` and `search` mode extraction.

Format

A tibble with 1 column:

text

A column wiht each row containing a sentence from the excerpt

Source

https://www.congress.gov/119/chrg/CHRG-119hhrg62658/CHRG-119hhrg62658.pdf

Details

An excerpt of text converted to a tibble from a 119 Congress hearing titled "Embedded Threats: Foreign Ownership, Hidden Hardware, and Licensing Failures in America’s Transportation Systems"